Add listingDecodor features
