+++ src/api/buildListingEndpoint/models.ts
@@ -1,5 +1,3 @@
-/* eslint-disable camelcase */
-
 export interface SearchInput {
   searchValue?: string;
   searchOptions?: Array<string>;
@@ -37,17 +35,4 @@ type Value = string | number | OrSearchParam | AndSearchParam | Sort;
 export interface Param {
   name: string;
   value?: Value;
-}
-
-export interface ListingMeta {
-  page: number;
-  limit: number;
-  search: AndSearchParam | OrSearchParam;
-  sort_by: Sort;
-  total: number;
-}
-
-export interface Listing<TEntity> {
-  result: Array<TEntity>;
-  meta: ListingMeta;
 }
