+++ src/api/models.ts
@@ -1,0 +1,10 @@
+export interface ListingMeta {
+  page: number;
+  limit: number;
+  total: number;
+}
+
+export interface Listing<TEntity> {
+  result: Array<TEntity>;
+  meta: ListingMeta;
+}
