+++ src/index.ts
@@ -102,6 +102,7 @@ export { default as RightPanel } from './RightPanel';
 export { default as StatusChip } from './StatusChip';
 export { SeverityCode, getStatusColors } from './StatusChip';
 
+export { Listing as ListingModel } from './api/models';
 export { default as useCancelTokenSource } from './api/useCancelTokenSource';
 export { getData, postData, putData, deleteData } from './api';
 export { default as useRequest } from './api/useRequest';
@@ -111,7 +112,7 @@ export {
   SearchInput,
   SearchObject,
   Param,
-  Listing as ListingModel,
 } from './api/buildListingEndpoint/models';
+export { default as buildListingDecoder } from './api/buildListingDecoder';
 
 export { default as copyToClipboard } from './utils/copy';
